pyproject.toml
[tool.poetry]
name = "vostoktelekom"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
python-jose = "^3.3.0"
passlib = "^1.7.4"
aiosqlite = "^0.20.0"
aiomysql = "^0.2.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
